Revert "WIP restructuring"
This reverts commit 28d49d1a6b4b35c89e1c4db988e745dc5ff5f515.
