--- .github/workflows/build-and-test.yml
@@ -1,12 +1,0 @@
-name: Build and Test .NET
-
-on:
-  pull_request_target:
-
-jobs:
-  build-and-test:
-    name: Run Shared Build and Test Workflow
-    uses: IntelliTect-Samples/EWU-CSCD371-2024-Winter/.github/workflows/build-and-test.yml@main
-    with:
-      check-formatting: false
-      add-analyzers: false
