+++ GenericsHomework.sln
@@ -3,12 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.8.34525.116
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GenericsHomework", "GenericsHomework\GenericsHomework.csproj", "{DCB98F54-AF50-4C72-9F22-FBBD0F6B1071}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GenericsHomework", "GenericsHomework\GenericsHomework.csproj", "{6991151A-B6D5-436B-8D2D-807252F2FF0B}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "GenericsHomework.Tests", "GenericsHomework.Tests", "{CE36A0B6-C84C-4E5F-8A40-DDC586F7CA69}"
-	ProjectSection(SolutionItems) = preProject
-		GenericsHomework.Tests\NodeTests.cs = GenericsHomework.Tests\NodeTests.cs
-	EndProjectSection
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GenericsHomeworkTests", "GenericsHomeworkTests\GenericsHomeworkTests.csproj", "{A7DD1BE0-202F-43BD-9141-9F95F65F339C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -16,15 +13,19 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{DCB98F54-AF50-4C72-9F22-FBBD0F6B1071}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DCB98F54-AF50-4C72-9F22-FBBD0F6B1071}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DCB98F54-AF50-4C72-9F22-FBBD0F6B1071}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DCB98F54-AF50-4C72-9F22-FBBD0F6B1071}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6991151A-B6D5-436B-8D2D-807252F2FF0B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6991151A-B6D5-436B-8D2D-807252F2FF0B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6991151A-B6D5-436B-8D2D-807252F2FF0B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6991151A-B6D5-436B-8D2D-807252F2FF0B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A7DD1BE0-202F-43BD-9141-9F95F65F339C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A7DD1BE0-202F-43BD-9141-9F95F65F339C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A7DD1BE0-202F-43BD-9141-9F95F65F339C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A7DD1BE0-202F-43BD-9141-9F95F65F339C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {E39DFB28-1846-4F0D-82C8-205D37B0F43F}
+		SolutionGuid = {263AE434-2EF2-409C-99BB-CAD20B2C2F45}
 	EndGlobalSection
 EndGlobal
