+++ GenericsHomework/GenericsHomeworkTests/GlobalUsings.cs
@@ -1,0 +1,1 @@
+global using Xunit;
